urban-farming-oasis-model/requirements.txt
Flask==2.1.0
numpy==1.24.1
openai==1.2.2
Pillow==10.1.0
tensorflow==2.14.0
tensorflow_gpu==2.8.0
